Derive breadcrumb content from navigation
